+++ change_log.md
@@ -1,5 +1,10 @@
 # Changelog
 
+**2.3.3** *30th June 2025*
+
+- Bugfix: sequence lengths are now calculated correctly when running `hybpiper stats` with the sequence type `supercontig`. See [issue#169](https://github.com/mossmatters/HybPiper/issues/169). Note that the GenesAt*pct columns in the stats report will not be accurate, as supercontigs can contain introns as well as coding exon sequence.
+- Fix: running `hybpiper stats` on compressed (`*.tar.gz`) sample folders with large `'BLASTX` output files now requires mich less memory.
+
 **2.3.2** *8th January 2025*
 
 - Bugfix: allow `hybpiper stats` to be run when gene names in the target file contain a dot (e.g. `taxon1-gene001.01`). See [issue#164](https://github.com/mossmatters/HybPiper/issues/164).